+++ src/web/assets/cp/src/js/TableElementIndexView.js
@@ -21,6 +21,9 @@ Craft.TableElementIndexView = Craft.BaseElementIndexView.extend({
 
   initialSerializedValue: null,
 
+  stickyScrollbar: null,
+  stickyScrollbarObserver: null,
+
   getElementContainer: function () {
     // Save a reference to the table
     this.$table = this.$container.find('table:first');
@@ -648,6 +651,13 @@ Craft.TableElementIndexView = Craft.BaseElementIndexView.extend({
       this.$cancelBtn.remove();
     }
 
+    if (this.stickyScrollbar) {
+      this.stickyScrollbar.remove();
+    }
+    if (this.stickyScrollbarObserver) {
+      this.stickyScrollbarObserver.disconnect();
+    }
+
     if (this._broadcastListener) {
       Craft.messageReceiver.removeEventListener(
         'message',
@@ -669,16 +679,16 @@ Craft.TableElementIndexView = Craft.BaseElementIndexView.extend({
       return;
     }
 
-    const stickyScrollbar = document.createElement('craft-proxy-scrollbar');
-    stickyScrollbar.setAttribute('scroller', '.tablepane');
-    stickyScrollbar.setAttribute('content', '.tablepane > table');
+    this.stickyScrollbar = document.createElement('craft-proxy-scrollbar');
+    this.stickyScrollbar.setAttribute('scroller', '.tablepane');
+    this.stickyScrollbar.setAttribute('content', '.tablepane > table');
 
-    stickyScrollbar.style.bottom = `${
+    this.stickyScrollbar.style.bottom = `${
       footer.getBoundingClientRect().height + 2
     }px`;
 
-    let $scrollbar = $(stickyScrollbar);
-    const observer = new IntersectionObserver(
+    let $scrollbar = $(this.stickyScrollbar);
+    this.stickyScrollbarObserver = new IntersectionObserver(
       ([ev]) => {
         if (ev.intersectionRatio < 1) {
           $scrollbar.insertAfter(this.$container);
@@ -691,6 +701,6 @@ Craft.TableElementIndexView = Craft.BaseElementIndexView.extend({
         threshold: [1],
       }
     );
-    observer.observe(footer);
+    this.stickyScrollbarObserver.observe(footer);
   },
 });